+++ code/disco-graph/ui/src/app/app-routing.module.ts
@@ -1,12 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DocumentComponent } from './components/document/document.component';
-import { SearchComponent } from './components/search/search.component';
 import { PageNotFoundComponent } from './components/util/page-not-found/page-not-found.component';
 
 const routes: Routes = [
-  {path: 'search', component: SearchComponent},
-  {path: 'document', component: DocumentComponent},
   {path: '', redirectTo: '/search', pathMatch: 'full'},
   {path: '**', component: PageNotFoundComponent}
 ]
