+++ code/disco-graph/ui/src/app/components/util/page-not-found/page-not-found.component.html
@@ -1,4 +1,4 @@
 <div class="centered-text">
     <h1>Page not found!</h1>
-    <p>The requested page is not available.</p>
+    <p>This page does not exist.</p>
 </div>
