+++ code/disco-graph/ui/src/app/document/document-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { SearchComponent } from './search/search.component';
+import { DocumentComponent } from './document/document.component';
 
 const routes: Routes = [
-  {path: 'search', component: SearchComponent},
+  {path: 'document/:id', component: DocumentComponent},
 ];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class SearchRoutingModule { }
+export class DocumentRoutingModule { }
