+++ code/disco-graph/ui/src/app/document/document/document.component.ts
