+++ code/disco-graph/ui/src/app/search/filters/attribute-filter/attribute-filter.component.scss
