+++ code/disco-graph/ui/src/app/search/filters/filter-pane/filter-pane.component.ts
