+++ code/disco-graph/ui/src/app/search/filters/years-filter/years-filter.component.scss
