+++ code/disco-graph/ui/src/app/search/input/keyword-input/keyword-input.component.scss
