+++ code/disco-graph/ui/src/app/search/result/result-item/result-item.component.scss
