+++ code/disco-graph/ui/src/app/search/result/results/results.component.html
