+++ code/disco-graph/ui/src/app/search/result/results/results.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../node_modules/nord/src/sass/nord.scss";
+@import "../../../../../node_modules/nord/src/sass/nord.scss";
 .mock-result-card {
     margin: auto;
     margin-bottom: 10px;
