+++ code/disco-graph/ui/src/app/search/result/results/results.component.spec.ts
