+++ code/disco-graph/ui/src/app/search/search/search.component.ts
