+++ code/disco-graph/ui/src/app/search/suggestion/keyword-suggestion/keyword-suggestion.component.ts
